// js/stat-evaluation-player/src/generated/PossessionEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

/**
 * A team-possession span.
 */
export type PossessionEvent = { time: number, frame: number, end_time: number, end_frame: number, active: boolean, duration: number, possession_state: string, player_id: RemoteIdTs | null, };